Move full_name method to decorator
Add draper dependency
Generate draper decorator for Student
Move full_name method to decorator
Move full_name model specs to student_decorator_spec
Add calls to decorate in Student controller
