+++ my-sms/Gemfile
@@ -7,6 +7,9 @@ gem 'rails', '~> 3.2.0'
 
 gem 'mysql2', '0.3.20'
 
+# rails gems
+gem 'draper'
+
 group :development do
   gem 'rubocop'
   gem 'spring-commands-rspec'
